requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
playwright
beautifulsoup4
lxml
pypdf
python-docx
markdown
trafilatura
openai
tiktoken
python-dotenv
httpx
jinja2
python-multipart
celery
redis
streamlit
pytest
black
isort
flake8
